Addressing recent change in ZF2
Seems that newer ZF2 no longer respect “Break chain on failure” when
the input filter is returned by an element specification.  You often
need CAPTCHA to validate first, and break if it failed.  So, add the
CAPTCHA validator directly to your InputFilter, and add it first with
the break chain specification like so:

```
class CaptchaPinInputFilter extends InputFilter
{
    const RECAPTCHA = 'g-recaptcha-response';

    public function init()
    {
        $this->add([
            'name'       => self::RECAPTCHA,
            'required'   => true,
            'messages' => array( _( "Please complete the anti-robot
check!" ) ),
        ]);

        $this->get( self::RECAPTCHA )->setBreakOnFailure( true );

        $this->add([
            'name' => 'pin',
            'required' => true,
            'filters'  => [
                ['name' => 'arrayblock'],
                ['name' => 'alnum'],
                ['name' => 'stringtrim'],
            ],
            'validators' => [
                array( 'name' => '\LDP\Form\Validator\PinAvailable' )
            ],
        ]);
    }
}
```